gzip tests: do not run alloc test with -race
Skip the allocation tests when running with the race detector, because
the race detector allocates a lot of memory.

Rename the unreliable tests file.
